Revert "Change Parcial"

+++ nds-service-client-impl/src/main/java/br/com/abril/nds/service/impl/ParciaisServiceImpl.java
@@ -188,13 +188,12 @@ public class ParciaisServiceImpl implements ParciaisService{
 
 				for (PeriodoLancamentoParcial item : lancamentoParcial.getPeriodos()) {
 
-					//TODO Ajuste alterações PARCIAIS
-					/*if( Arrays.asList(StatusLancamento.PLANEJADO, 
+					if( Arrays.asList(StatusLancamento.PLANEJADO, 
 									  StatusLancamento.CONFIRMADO)
 									  .contains(item.getLancamento().getStatus())){
 						
 						periodoLancamentoParcialRepository.remover(item);
-					}*/
+					}
 				}
 				
 				return true;
@@ -217,8 +216,7 @@ public class ParciaisServiceImpl implements ParciaisService{
 	private PeriodoLancamentoParcial gerarPeriodoParcial(Lancamento lancamento, LancamentoParcial lancamentoParcial) {
 		
 		PeriodoLancamentoParcial periodo = new PeriodoLancamentoParcial();
-		//TODO Ajuste alterações PARCIAIS
-		//periodo.setLancamento(lancamento);
+		periodo.setLancamento(lancamento);
 		periodo.setLancamentoParcial(lancamentoParcial);
 		periodo.setTipo(TipoLancamentoParcial.PARCIAL);
 		periodo.setStatus(StatusLancamentoParcial.PROJETADO);
